remove unneeded copies from ds6-safety.cpp fixes wrappers/python/c_files.cpp
